changed input of stepcooling from real to integer
the old step cooling used to check if two reals were equal to another to activate. checking if 2 reals are equal is finicky. but checking to see if 2 integers are equal is easy and unlikely to cause errors. could be a nasty bug to fix if something goes wrong later. the code now activates step cooling if two integers are equal. thanks mike!